Move options handling to dedicated file
Similar to https://github.com/gardener/gardener/blob/master/cmd/gardener-controller-manager/app/options.go
