+++ cmd/gardener-admission-controller/app/options.go
@@ -1,0 +1,75 @@
+// Copyright (c) 2022 SAP SE or an SAP affiliate company. All rights reserved. This file is licensed under the Apache Software License, v. 2 except as noted otherwise in the LICENSE file
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//      http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package app
+
+import (
+	"fmt"
+	"os"
+
+	"github.com/spf13/pflag"
+	"k8s.io/apimachinery/pkg/runtime"
+	"k8s.io/apimachinery/pkg/runtime/serializer"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+
+	"github.com/gardener/gardener/pkg/admissioncontroller/apis/config"
+	configv1alpha1 "github.com/gardener/gardener/pkg/admissioncontroller/apis/config/v1alpha1"
+	configvalidation "github.com/gardener/gardener/pkg/admissioncontroller/apis/config/validation"
+)
+
+var configDecoder runtime.Decoder
+
+func init() {
+	configScheme := runtime.NewScheme()
+	schemeBuilder := runtime.NewSchemeBuilder(
+		config.AddToScheme,
+		configv1alpha1.AddToScheme,
+	)
+	utilruntime.Must(schemeBuilder.AddToScheme(configScheme))
+	configDecoder = serializer.NewCodecFactory(configScheme).UniversalDecoder()
+}
+
+type options struct {
+	configFile string
+	config     *config.AdmissionControllerConfiguration
+}
+
+func (o *options) addFlags(fs *pflag.FlagSet) {
+	fs.StringVar(&o.configFile, "config", o.configFile, "Path to configuration file.")
+}
+
+func (o *options) complete() error {
+	if len(o.configFile) == 0 {
+		return fmt.Errorf("missing config file")
+	}
+
+	data, err := os.ReadFile(o.configFile)
+	if err != nil {
+		return fmt.Errorf("error reading config file: %w", err)
+	}
+
+	o.config = &config.AdmissionControllerConfiguration{}
+	if err = runtime.DecodeInto(configDecoder, data, o.config); err != nil {
+		return fmt.Errorf("error decoding config: %w", err)
+	}
+
+	return nil
+}
+
+func (o *options) validate() error {
+	if errs := configvalidation.ValidateAdmissionControllerConfiguration(o.config); len(errs) > 0 {
+		return errs.ToAggregate()
+	}
+	return nil
+}
